refactor: refactor hd wallet and address
